requirements.txt
psycopg2-binary==2.9.9
PyYAML==6.0.1
fastapi==0.115.0
uvicorn==0.30.6
